Initial commit of XmlMapper module
This is the initial commit of the XmlMapper module. This code is being
broken out from the trlinkin/puppet-weblogic repository since it is very
general purpose and usable beyond Weblogic configuration.

This commit takes the code and picks up there commit 525c772 to
trlinkin/puppet-weblogic left off. This commit also updates the names in
the files to reflect their new names.