Add support to manage the XML Declaration
Without this commit, there was no way to manage the xml declaration in
the xmlmapper.

This commit adds a new special component 'xmldeclcomponent' that is
used to represent an xml declaration. It also adds a helper method that
will create methods to manage the existence of the xmldecl as described
in the block passed to 'xmldecl'.