Depend on resource to provide document path
Without this commit, the provider was looking for the developer to
indicate which variable in resource contained the document_path. This is
not really a sound interface between the resource and provider and makes
the code very inflexible.

This commit changes the relationship between resource and provider. This
change is not backwards compatible but does enable smarter interactions
in the future. With this change, the provider now expects the resource
to implement a `document_path` method. This method on the resource
should return a valid file path. The resource itself remains responsible
for validation of the file path.