Refactor access to 'resource' object from provider
Without this commit, the provider was accessing the instance variable
`@resource` directly. This goes against good practice in general, but
is made worse by the fact that I switching between the resource method
and the instance variable for no reason.

This commit refactor all access to the resource object to be via the
accessor function.